+++ CHANGELOG.md
@@ -37,6 +37,10 @@ Section Order:
 ### Security
 -->
 
+### Changed
+
+- `Settings` moved into the `Sources` folder, to keep the root folder clean
+
 ### [1.4.0] - 2024-12-09
 
 #### Fixed